+++ src/brasil/gov/portlets/config.py
@@ -13,7 +13,8 @@ class HiddenProducts(object):
     def getNonInstallableProducts(self):
         return [
             u'brasil.gov.portlets.upgrades.v1000',
-            u'brasil.gov.portlets.upgrades.v1001'
+            u'brasil.gov.portlets.upgrades.v1001',
+            u'brasil.gov.portlets.upgrades.v1002',
         ]
 
 
@@ -23,6 +24,8 @@ class HiddenProfiles(object):
     def getNonInstallableProfiles(self):
         return [
             u'brasil.gov.portlets:uninstall',
+            u'brasil.gov.portlets:testfixture',
             u'brasil.gov.portlets.upgrades.v1000:default',
-            u'brasil.gov.portlets.upgrades.v1001:default'
+            u'brasil.gov.portlets.upgrades.v1001:default',
+            u'brasil.gov.portlets.upgrades.v1002:default',
         ]